requirements.txt
numpy==2.1.3
openai==0.28.1
pandas==2.2.3
python-dotenv==1.0.1
vl-convert-python
